+++ CHANGES.md
@@ -1,8 +1,11 @@
 # CHANGES
 
-## NEXT
+## 5.2.0
 
-* Expose option for "fastcollate" to allow proxy for <=5.0.5 processing methods.
+* Expose option for "legacy" to allow for <=5.0.5 processing methods.
+  * `bamtofastq` when pulling reads from BAM/CRAM input.
+  * `bammarkduplicates2` for duplicate marking.
+  * Affects `bwa_mem.pl` and `merge_or_mark.pl`
 
 ## 5.1.0